Turn the CEF plugin into a builtin
