+++ .github/workflows/vast.yaml
@@ -940,9 +940,6 @@ jobs:
           - name: PCAP
             target: pcap
             path: plugins/pcap
-          - name: CEF
-            target: cef
-            path: plugins/cef
           - name: Web
             target: web
             path: plugins/web
