+++ changelog/unreleased/features/2216--cef-reader.md
@@ -1,0 +1,2 @@
+The `cef` import format allows for reading events in the Common Event Format
+(CEF) via `vast import cef < cef.log`.
