--- plugins/cef/CHANGELOG.md
@@ -1,7 +1,0 @@
-# Changelog
-
-This changelog documents all notable changes to the CEF plugin for VAST.
-
-## v0.1
-
-This is the first official release.
