--- plugins/cef/integration/tests.yaml
@@ -1,28 +1,0 @@
-config-file: vast.yaml
-
-fixtures:
-  ServerTester:
-    enter: | # python
-      node = Server(self.cmd,
-                    ['-e', f'127.0.0.1:{VAST_PORT}', '-i', 'node', 'start'],
-                    work_dir, name='node', port=VAST_PORT,
-                    config_file=self.config_file)
-      cmd += ['-e', f'127.0.0.1:{VAST_PORT}']
-
-    exit: | # python
-      node.stop()
-
-tests:
-  CEF:
-    tags: [server, import-export, cef]
-    fixture: ServerTester
-    steps:
-      - command: import -b cef
-        input: data/cef/cynet.log
-      - command: import -b cef
-        input: data/cef/checkpoint.log
-      - command: import -b cef
-        input: data/cef/forcepoint.log
-      - command: export json 'cef_version >= 0 && device_vendor == "Cynet"'
-      - command: export json '172.31.5.93'
-      - command: export json 'act == /Accept|Bypass/'
