--- plugins/cef/integration/vast.yaml
@@ -1,3 +1,0 @@
-vast:
-  plugins:
-    - cef
