+++ vast/integration/data/cef/checkpoint.log
