+++ vast/integration/vast_integration_suite.yaml
@@ -841,3 +841,17 @@ tests:
         transformation: jq -ers 'length'
       - command: export --max-events=1 json 'head 0'
       - command: export json 'head 0'
+
+  CEF:
+    tags: [server, import-export, cef]
+    fixture: ServerTester
+    steps:
+      - command: import -b cef
+        input: data/cef/cynet.log
+      - command: import -b cef
+        input: data/cef/checkpoint.log
+      - command: import -b cef
+        input: data/cef/forcepoint.log
+      - command: export json 'cef_version >= 0 && device_vendor == "Cynet"'
+      - command: export json '172.31.5.93'
+      - command: export json 'act == /Accept|Bypass/'
